+++ test/server.test.ts
@@ -5,7 +5,7 @@
 import * as child_process from "child_process";
 import * as assert from "assert";
 
-import { TextDocumentSyncKind, MarkupKind } from 'vscode-languageserver';
+import { TextDocumentSyncKind, MarkupKind, SymbolKind } from 'vscode-languageserver';
 import { CommandIds } from 'dockerfile-language-service';
 import { ValidationCode } from 'dockerfile-utils';
 
@@ -374,5 +374,49 @@ describe("Dockerfile LSP Tests", function() {
 			}
 		};
 		lspProcess.on("message", listener218);
+	});
+
+	it("issue #223", function (finished) {
+		this.timeout(5000);
+		let document = {
+			languageId: "dockerfile",
+			version: 1,
+			uri: "uri://dockerfile/223.txt",
+			text: "FROM node\nMAINTAINER deprecated"
+		};
+		sendNotification("textDocument/didOpen", {
+			textDocument: document
+		});
+
+		let id = sendRequest("textDocument/documentSymbol", {
+			textDocument: {
+				uri: document.uri
+			}
+		});
+		let first = true;
+		const listener223 = (json) => {
+			if (json.id === id) {
+				assert.equal(json.result[0].name, "FROM");
+				assert.equal(json.result[0].location.uri, document.uri);
+				assert.equal(json.result[0].location.range.start.line, 0);
+				assert.equal(json.result[0].location.range.start.character, 0);
+				assert.equal(json.result[0].location.range.end.line, 0);
+				assert.equal(json.result[0].location.range.end.character, 4);
+				assert.equal(json.result[0].kind, SymbolKind.Function);
+				assert.strictEqual(json.result[0].deprecated, undefined);
+
+				assert.equal(json.result[1].name, "MAINTAINER");
+				assert.equal(json.result[1].location.uri, document.uri);
+				assert.equal(json.result[1].location.range.start.line, 1);
+				assert.equal(json.result[1].location.range.start.character, 0);
+				assert.equal(json.result[1].location.range.end.line, 1);
+				assert.equal(json.result[1].location.range.end.character, 10);
+				assert.equal(json.result[1].kind, SymbolKind.Function);
+				assert.equal(json.result[1].deprecated, true);
+				lspProcess.removeListener("message", listener223);
+				finished();
+			}
+		};
+		lspProcess.on("message", listener223);
 	});
 });